tpm2: Replace MAX_SM4_KEY_BITS and MAX_CAMELLIA_KEY_BITS with 128
Rev 155 code defines MAX_SM4_KEY_BITS = 0 and MAX_CAMELLIA_KEY_BITS = 0
since both algorithms are not enabled. To be able to read old state
we now replace the usage of the #defines with hard coded values that
have no further meaning. If we ever activate the algoritms, we have to
extend the state.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>